Change Renderer from class to namespace
Since the renderer holds a lot of global state, it doesn't make sense to
have it be a class. Instead, putting the public interface in a namespace
and the private interface in an unnamed namespace does the job just as
well. The public header file looks much cleaner because of that as well.